refactor: promote passive hyper skill into jobs
