+++ simaple/data/jobs/__init__.py
@@ -1,5 +1,6 @@
-from simaple.data.jobs._builtin import get_skill_profile
+from simaple.data.jobs.builtin import get_skill_profile, get_every_hyper_skills
 
 __all__ = [
     "get_skill_profile",
+    "get_every_hyper_skills",
 ]
