+++ simaple/data/jobs/builtin.py
@@ -10,6 +10,7 @@ from simaple.spec.loadable import (  # pylint:disable=unused-import
 from simaple.spec.loader import SpecBasedLoader
 from simaple.spec.repository import DirectorySpecRepository
 from simaple.data.jobs.definitions.skill_profile import SkillProfile
+from simaple.data.jobs.definitions import PassiveHyperskillInterface
 
 
 
@@ -23,4 +24,11 @@ def get_skill_profile(jobtype: JobType) -> SkillProfile:
         loader.load(
             query={"group": jobtype.value, "kind": "SkillProfile"},
         ),
+    )
+
+def get_every_hyper_skills(group: str) -> list[PassiveHyperskillInterface]:
+    repository = DirectorySpecRepository(str(Path(__file__).parent / "resources"))
+    loader = SpecBasedLoader(repository)
+    return loader.load_all(
+        query={"group": group, "kind": "PassiveHyperskill"},
     )
