+++ simaple/data/jobs/definitions/passive_hyper_skill.py
