+++ simaple/data/jobs/resources/passive_hyper_skill/adele.yaml
