+++ simaple/data/jobs/resources/passive_hyper_skill/archmagefb.yaml
