+++ simaple/data/jobs/resources/passive_hyper_skill/mechanic.yaml
