+++ simaple/data/jobs/resources/passive_hyper_skill/windbreaker.yaml
